requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
numpy>=1.24
pandas>=1.5
opencv-python>=4.8
pillow>=9.0
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.64
pytest>=7.0
mlflow>=2.12
